requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot
python-dotenv
pydantic
langchain-core
langgraph
langgraph-checkpoint
openai
groq
instructor
psycopg2-binary
tavily-python
twilio
aiohttp
aiohttp-retry
requests
python-multipart
google-generativeai    
numpy
mcp
sqlalchemy                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              